+++ skysail.product.demo/src/io/skysail/product/demo/DemoRootApplication.java
@@ -1,7 +1,5 @@
 package io.skysail.product.demo;
 
-import org.osgi.service.event.EventAdmin;
-
 import io.skysail.server.app.ApplicationProvider;
 import io.skysail.server.app.SkysailApplication;
 import io.skysail.server.restlet.RouteBuilder;
@@ -12,11 +10,6 @@ public class DemoRootApplication extends SkysailApplication implements Applicati
         super("demoproduct", null);
     }
     
-    @Override
-    public EventAdmin getEventAdmin() {
-        return null;
-    }
-
     @Override
     protected void attach() {
         router.attach(new RouteBuilder("", PublicResource.class).noAuthenticationNeeded());