+++ skysail.server.app.demo/src/io/skysail/server/app/demo/DemoApplication.java
@@ -107,9 +107,4 @@ public class DemoApplication extends SkysailApplication implements ApplicationPr
                 
     }
 
-    @Override
-	public EventAdmin getEventAdmin() {
-        return eventAdmin;
-    }
-	
 }