+++ skysail.server.um.httpbasic/src/io/skysail/server/um/httpbasic/app/HttpBasicUmApplication.java
@@ -1,0 +1,8 @@
+package io.skysail.server.um.httpbasic.app;
+
+import io.skysail.server.app.SkysailApplication;
+
+public class HttpBasicUmApplication extends SkysailApplication {
+
+
+}