+++ skysail.server/test/io/skysail/server/model/test/BreadcrumbsTest.java
@@ -3,18 +3,20 @@ package io.skysail.server.model.test;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.List;
 
-import org.junit.*;
+import org.junit.Before;
+import org.junit.Test;
 import org.mockito.Mockito;
-import org.osgi.service.event.EventAdmin;
 import org.restlet.Request;
 import org.restlet.data.Reference;
 import org.restlet.routing.Route;
 import org.restlet.util.RouteList;
 
 import io.skysail.server.app.SkysailApplication;
-import io.skysail.server.model.*;
+import io.skysail.server.model.Breadcrumb;
+import io.skysail.server.model.Breadcrumbs;
 import io.skysail.server.restlet.resources.SkysailServerResource;
 
 public class BreadcrumbsTest {
@@ -34,15 +36,9 @@ public class BreadcrumbsTest {
         Mockito.when(resource.getReference()).thenReturn(reference);
         Mockito.when(resource.getRequest()).thenReturn(request);
         Mockito.when(resource.getApplication()).thenReturn(new SkysailApplication("appName") {
-
             @Override
             public RouteList getRoutes() {
                 return routes;
-            }
-
-            @Override
-            public EventAdmin getEventAdmin() {
-                return null;
             }
         });
     }